wrote friends functionality & misc.
users can now have friends which are stored in an array with references to user id's. The routes to add these friends will attempt to check if the added id is an actual user id, aborting the process if it does not. Also added other properties, such as posts now containing the username of its author. Right now, this process is not automatic and the request body can insert any string for post username.